pyproject.toml
[project]
name = "csa-lab-4"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.4",
    "pyyaml>=6.0.3",
]
